+++ openstack/rds/v3/datastores/List.go
@@ -5,12 +5,26 @@ import (
 	"github.com/opentelekomcloud/gophertelekomcloud/pagination"
 )
 
+func List(client *golangsdk.ServiceClient, databasesname string) pagination.Pager {
+	url := client.ServiceURL("datastores", databasesname)
+
+	pageRdsList := pagination.NewPager(client, url, func(r pagination.PageResult) pagination.Page {
+		return DataStoresPage{pagination.SinglePageBase(r)}
+	})
+
+	rdsheader := map[string]string{"Content-Type": "application/json"}
+	pageRdsList.Headers = rdsheader
+	return pageRdsList
+}
+
 type DataStoresResult struct {
 	golangsdk.Result
 }
+
 type DataStores struct {
 	DataStores []dataStores `json:"dataStores" `
 }
+
 type dataStores struct {
 	Id   string `json:"id" `
 	Name string `json:"name"`
