--- openstack/rds/v3/datastores/urls.go
@@ -1,7 +1,0 @@
-package datastores
-
-import "github.com/opentelekomcloud/gophertelekomcloud"
-
-func listURL(sc *golangsdk.ServiceClient, databasename string) string {
-	return sc.ServiceURL("datastores", databasename)
-}
